[IMP] Blocked locations handling core code
Stock location extended with 2 fields:
- u_blocked: boolean that determines if the locaiton is blocked
- u_blocked_reason: text with the reason that has been blocked

Stock location form view updated to show u_blocked field, when it is
set to true the reason text field is shown and when it is set to false
it is hidden and erased.

Filter by blocked locations added.

When a location is blocked, its containing quants are not available
for reservation.

Blocked locations cannot be used as source or destination location
for pickings, moves, move lines, inventories or inventory lines.

Stock inventories cannot be validated if the location has been blocked
after being created.